fixes for tags
